backend/requirements.txt
google-adk[full]>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
websockets>=12.0